requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
supabase
python-dotenv
